fix: comparison of node names are case insensitive
when the parser in AngleSharp does not know the element type, it will store the name of the node/element in the NodeName property in upper case. This can lead to false negative results. E.g. if one parses "<svg><path></path></svg>" and selects the `<path />` and want to compare that to a "<path>" element parsed alone, the second will be treated as an unknown element, since it was not parsed in the context of a "<svg>" element. The two elements will have the same name though, just in different cases and should thus be equal.